backend/requirements.txt
ftfy
python-dotenv
fastapi
uvicorn[standard]
requests
sqlalchemy
mysqlclient
python-multipart
alembic
pytest
httpx